tweets and media queries
here i added the final layout element which was the tweets and then made the navbar and the aside sticky as it is with twitter. I have then created a new file with media queries making the site fully responsive down to the smallest screen sizes. i have also built 3 versions of the navbar for the respective viewports to be the same that twitter offers.